Rework the saved templates to use virtual list, too.
