+++ assets/src/edit-story/components/library/panes/pageTemplates/templateList.js
@@ -20,8 +20,6 @@
 import { useCallback, useMemo, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { useVirtual } from 'react-virtual';
-import styled from 'styled-components';
-import { v4 as uuidv4 } from 'uuid';
 import { __ } from '@web-stories-wp/i18n';
 
 /**
@@ -36,42 +34,13 @@ import {
   PANEL_GRID_ROW_GAP,
   VirtualizedWrapper,
 } from '../shared/virtualizedPanelGrid';
-import {
-  Headline,
-  Text,
-  THEME_CONSTANTS,
-  Toggle,
-} from '../../../../../design-system';
 import PageTemplate from './pageTemplate';
 import ConfirmPageTemplateDialog from './confirmPageTemplateDialog';
 import useTemplateActions from './useTemplateActions';
 
-const ActionRow = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin: 8px 16px 22px 16px;
-`;
-
-const TemplatesToggle = styled.div`
-  display: flex;
-
-  label {
-    cursor: pointer;
-    margin: auto 12px;
-    color: ${({ theme }) => theme.colors.fg.secondary};
-  }
-`;
-
-function DefaultTemplatesList({
-  onToggleClick,
-  pages,
-  parentRef,
-  showTemplateImages,
-  pageSize,
-}) {
+function TemplateList({ pages, parentRef, pageSize }) {
   const containerRef = useRef();
   const pageRefs = useRef({});
-  const toggleId = useMemo(() => `toggle_page_templates_${uuidv4()}`, []);
 
   const pageIds = useMemo(() => pages.map((page) => page.id), [pages]);
 
@@ -132,29 +101,6 @@ function DefaultTemplatesList({
         height: pageSize.height,
       }}
     >
-      <ActionRow>
-        <Headline
-          as="h3"
-          size={THEME_CONSTANTS.TYPOGRAPHY.PRESET_SIZES.XXX_SMALL}
-        >
-          {__('Templates', 'web-stories')}
-        </Headline>
-        <TemplatesToggle>
-          <Text
-            as="label"
-            htmlFor={toggleId}
-            size={THEME_CONSTANTS.TYPOGRAPHY.PRESET_SIZES.SMALL}
-          >
-            {__('Show Images', 'web-stories')}
-          </Text>
-          <Toggle
-            id={toggleId}
-            name={toggleId}
-            checked={showTemplateImages}
-            onChange={onToggleClick}
-          />
-        </TemplatesToggle>
-      </ActionRow>
       <VirtualizedWrapper height={rowVirtualizer.totalSize}>
         <VirtualizedContainer
           height={rowVirtualizer.totalSize}
@@ -209,16 +155,14 @@ function DefaultTemplatesList({
   );
 }
 
-DefaultTemplatesList.propTypes = {
-  onToggleClick: PropTypes.func.isRequired,
+TemplateList.propTypes = {
   parentRef: PropTypes.object.isRequired,
   pages: PropTypes.arrayOf(
     PropTypes.shape({
       id: PropTypes.string.isRequired,
     })
   ),
-  showTemplateImages: PropTypes.bool,
   pageSize: PropTypes.object.isRequired,
 };
 
-export default DefaultTemplatesList;
+export default TemplateList;
